Removed ContributorCard test, isolated contributors array
