--- components/ContributorCard.test.js
@@ -1,21 +1,0 @@
-import * as React from 'react'
-import ContributorCard from './ContributorCard'
-import { render } from '@testing-library/react'
-
-describe('ContributorCard', () => {
-  test('Should render ContributorCard with user data', async () => {
-    const user = {
-      firstName: 'Alexis',
-      lastName: 'Ortiz Ojeda',
-      username: 'aortizoj',
-      linkedinUrl: 'https://www.linkedin.com/in/alexis-ortiz-ojeda/',
-      githubPRsUrl:
-        'https://github.com/garageScript/c0d3-app/pulls?q=is%3Apr+author%3Aaortizoj15',
-      codeSample: 'https://github.com/garageScript/c0d3-app/pull/67',
-      resume:
-        'https://drive.google.com/file/d/1HoMg9QHGH2qicG8bZ8hRWv7FT7GXjFB5/view'
-    }
-    const { container } = render(<ContributorCard user={user} />)
-    expect(container).toMatchSnapshot()
-  })
-})
