+++ contributors.js
@@ -1,0 +1,15 @@
+const contributors = [
+  {
+    firstName: 'Alexis',
+    lastName: 'Ortiz Ojeda',
+    username: 'aortizoj',
+    linkedinUrl: 'https://www.linkedin.com/in/alexis-ortiz-ojeda/',
+    githubPRsUrl:
+      'https://github.com/garageScript/c0d3-app/pulls?q=is%3Apr+author%3Aaortizoj15',
+    codeSample: 'https://github.com/garageScript/c0d3-app/pull/67',
+    resume:
+      'https://drive.google.com/file/d/1HoMg9QHGH2qicG8bZ8hRWv7FT7GXjFB5/view'
+  }
+]
+
+export default contributors
