+++ scss/contributorCard.scss
@@ -1,4 +1,4 @@
-.contributorCard_avatarContainer{
+.contributorCard__avatarContainer{
  height: 60px;
  width: 60px;
 }
